Add zero-pad support to CPU reference model

+++ Tensile/Source/client/source/ClientProblemFactory.cpp
@@ -119,8 +119,7 @@ namespace Tensile
                     if (zp.size() % 4 != 0)
                         throw std::runtime_error("zero-pad must contain tuples of 4 values");
                     for (int zi=0; zi<zp.size(); zi+=4) {
-                        rv.back().aZeroPad().push_back(ContractionProblem::ZeroPad({zp[zi+0],zp[zi+1],
-                                    static_cast<int>(zp[zi+2]), static_cast<int>(zp[zi+3])}));
+                        rv.back().aZeroPad().push_back(ContractionProblem::ZeroPad({zp[zi+0],zp[zi+1],zp[zi+2], zp[zi+3]}));
                     }
                 }
                 if (i<m_bZeroPads.size())
@@ -129,8 +128,7 @@ namespace Tensile
                     if (zp.size() % 4 != 0)
                         throw std::runtime_error("zero-pad must contain tuples of 4 values");
                     for (int zi=0; zi<zp.size(); zi+=4) {
-                        rv.back().bZeroPad().push_back(ContractionProblem::ZeroPad({zp[zi+0],zp[zi+1],
-                                    static_cast<int>(zp[zi+2]), static_cast<int>(zp[zi+3])}));
+                        rv.back().bZeroPad().push_back(ContractionProblem::ZeroPad({zp[zi+0],zp[zi+1],zp[zi+2], zp[zi+3]}));
                     }
                 }
                 rv.back().setHighPrecisionAccumulate(m_highPrecisionAccumulate);
